chore(TCOMP-681): move ElementParameterCreator to model.parameter package
